requirements.txt
numpy==1.26.4
opencv-python==4.8.0.76
mediapipe
flask
boto3
pillow
pytesseract